Implement fix for Google Auth problem
The problem was that the StorageClient.Create() was failing with
an authentication error. By trial and error, I found that removing
the Google.Cloud.Translation.V2 package from the library solve it.
It was likely conflicting with Google.Cloud.Storage.V1
I removed the file "translation.cs" from the library and put in it
the TranslationText project. I added  Google.Cloud.Translation.V2
to that project.